refactor(fe): extract useAiCheckForm hook and FormField component
Moves form state and submit logic into useAiCheckForm hook.
Extracts per-field rendering into reusable FormField component.

via [Happy](https://happy.engineering)
